add basic routing for onboarding
