+++ app.json
@@ -1,7 +1,8 @@
 {
   "expo": {
     "name": "Bitgesell Wallet",
-    "slug": "Bglwallet",
+    "slug": " bitgesell-wallet",
+    "scheme": "bitgesell-wallet",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
@@ -30,5 +31,8 @@
     "plugins": [
       "expo-router"
     ]
+  },
+  "web": {
+    "bundler": "metro"
   }
 }
