+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "bglwallet",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -11,24 +11,29 @@
   "dependencies": {
     "@expo-google-fonts/poppins": "^0.2.3",
     "expo": "~49.0.15",
-    "expo-constants": "~14.4.2",
-    "expo-linking": "~5.0.2",
+    "expo-constants": "^14.4.2",
+    "expo-linking": "^6.0.0",
     "expo-router": "^2.0.0",
     "expo-splash-screen": "~0.20.5",
-    "expo-status-bar": "~1.6.0",
+    "expo-status-bar": "^1.7.1",
     "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-native": "0.72.6",
     "react-native-app-intro-slider": "^4.0.4",
     "react-native-element-dropdown": "^2.10.0",
-    "react-native-gesture-handler": "~2.12.0",
-    "react-native-safe-area-context": "4.6.3",
-    "react-native-screens": "~3.22.0",
-    "react-native-svg": "13.9.0"
+    "react-native-gesture-handler": "^2.14.0",
+    "react-native-safe-area-context": "^4.8.1",
+    "react-native-screens": "^3.29.0",
+    "react-native-svg": "13.9.0",
+    "react-native-web": "~0.19.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
     "typescript": "^5.1.3"
+  },
+  "overrides": {
+    "react-refresh": "~0.14.0"
   },
   "private": true
 }
